Performance: Optimize `SitesList.markCollisions`
In this patch I have rewritten `markCollisions` with performance in
mind. Recently we were investigating freezes in Calypso (#19446) and
noticed that `SitesList.sync()` was getting called over and over again
and that this particular part of that process, a quadric operation in
terms of complexity, was potentially a big part of our CPU usage.

Although I could not reproduce the freezing behavior I went ahead and
tried to refactor this function so that it would remain linear with the
number of sites (well, it's `O( 2 * N )` now because it iterates twice).

The major change is storing the known Jetpack sites by URL in a mutable
map for quick retrieval. A quick pass can add in the jetpack sites then
a second pass can quickly determine if there are non-matching sites
sharing the same URL.

That is to say, we have replaced loop iteration with key lookup.

In my tests this was more than a magnitude in improvement, but overall
insignificant because I was only seeing this function take up 4-5ms on
average to run, plus it only ran a few times. (the function after these
changes was taking about 0.25ms)

If, however, we get into a loop and call this function a few hundred
times (as appears to have happened in a performance trace) then we could
end up shaving off a few hundred milliseconds of freeze time which
_would indeed_ be noticeable.